Clean up the documentation and public exports
The interrupt helper struct does not need to be public.

Add some documentation to a bunch of types.
