+++ src/legacy/mod.rs
@@ -1,6 +1,7 @@
 //! Legacy code.
 //!
-//! The code in here needs to be cleaned up and rewritten.
+//! The code in here needs to be cleaned up and rewritten to use
+//! the new core-based API.
 //!
 //! Once a replacement exists, these legacy modules may be removed
 //! in a major release!
